+++ src/publish.rs
@@ -1,0 +1,4 @@
+use crate::observer::Observer;
+
+pub trait Publish<Item, Err>: Observer<Item, Err> {
+}